+++ src/app/components/modals/detail-modal/detail-modal.component.html
