+++ src/app/components/modals/detail-modal/detail-modal.component.ts
